The merged build no longer rearranges 1,234 players on its first launch
Six things an adversarial pass found the merge would do to a live 2.2 GB world, fixed before
anyone runs it.

THE YOUTH SELF-HEAL MOVED PEOPLE. The world-data rebuild dropped every U21/U18 side while
leaving youth_teams_v1 at '1', so EnsureYouthTeamsOnLoad fired its full pass on first launch —
and EnsureYouthSide did two jobs in one: create the side, then MOVE every deep-squad player
aged in band out of the senior squad. Measured: 1,234 players out of 44 first teams, Manchester
United 92 to 30, inside a bare catch. Creation and seeding are separate now. A repair creates
the sides EMPTY and posts a letter saying so; only a career that has genuinely not begun
(no results, no inbox — the honest test, since the constructor writes the welcome letter at
the end of the first load) still seeds from the reserves. MovePlayer had the same fault: sending
one lad down could drag every eligible reserve with him.

THE OPENING-XI GUARD LOOKED AT ONE SLOT. Liverpool's slots 0-3 are four goalkeepers; the test
saw Alisson at slot 0 and passed. XiLooksUnarranged now counts keepers and category agreement
across the whole eleven.

TWO ID TRUNCATIONS ON EVERY LAUNCH. SessionXiRepair read squad_members.player_id with GetInt32
— 304,437 of 337,407 rows exceed Int32 — and produced half-applied slot swaps; SessionRoles
keyed its ability map on a truncated id, so roles compiled into Player.bin from an empty
dictionary. Both read longs now; SwapSlots is one transaction.

THE SCHEMA LIED ABOUT superseded_by. It exists on the live file only because a tool ALTERed
it in; the merged Repository queries it; schema.sql and Migrate knew nothing. A fresh or
in-memory database threw "no such column" on the first Players() call. Added to both.

THE CAREER COULD BE SILENTLY SWAPPED. CareerLoader caught every exception and returned null,
and the caller fell through to the sample career — so a locked file would have looked exactly
like "the update broke my save". The failure is logged and carried to the picker. DropHonoursTeamFk
no longer propagates.

AND THE COMPLAINT ITSELF. Merging alone did not answer it: the 336px rail is on both branches,
single-click still filled it, and neither new gesture had a visible trigger. Every roster row
now carries a ⋯ that opens the same menu as right-click, the rail carries "Open full profile",
and the screen says so in words. MlMenu.Attach is guarded, and shortlist mode can no longer
clear the one-record-per-human filter.

+++ src/ML.App/SessionRoles.cs
@@ -12,6 +12,8 @@ namespace ML.App;
 public sealed partial class Session
 {
     /// <summary>The team's chosen in-possession style (0-5), or null if unset.</summary>
+    // teamId and the style are legitimately int — the largest club id in the world is 4,032,815
+    // and a style is 0-5. Player ids are the ones that must be long; don't "tidy" these to match.
     private int? TeamStyleOf(int teamId)
     {
         try
@@ -76,8 +78,14 @@ public sealed partial class Session
         using var r = cmd.ExecuteReader();
         while (r.Read())
         {
-            var pid = r.GetInt32(0);
+            // THE BUG: this dictionary was keyed on GetInt32(player_id) — an unchecked truncation
+            // of the int64 column — while AssignRolesFor looks it up with the real long id. 8.27M
+            // of the world's 10.88M attribute rows are above Int32.MaxValue, and not one of the
+            // current club's 25 players survived the cast: every lookup missed, so both roles were
+            // picked from an EMPTY ability map and then compiled into Player.bin.
+            var pid = r.GetInt64(0);
             if (!map.TryGetValue(pid, out var m)) map[pid] = m = new Dictionary<string, int>();
+            // The attribute value is legitimately an int (0-95 across the world) — leave it.
             m[r.GetString(1)] = r.GetInt32(2);
         }
         return map;
